+++ src/main/java/com/mi_web/app/repositories/OrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-    List<Order> findByRestaurantId(Integer restaurantId);
-    List<Order> findByUserId(Integer userId);
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    List<Order> findByRestaurantId(Long restaurantId);
+    List<Order> findByUserId(Long userId);
 }
